Cmake extension

+++ shm/island.cpp
@@ -1,9 +1,9 @@
 #include "island.hpp"
 
-Island::Coordinates::Coordinates(int positionX, int positionY)
+Island::Coordinates::Coordinates(size_t positionX, size_t positionY)
     : positionX_(positionX), positionY_(positionY) {}
 
-Island::Island(int position_X, int position_Y)
+Island::Island(size_t position_X, size_t position_Y)
     : position_(Coordinates(position_X, position_Y)) {}
 
 bool Island::Coordinates::operator==(const Coordinates& rhs) const {
